refactor: move and update testability fixtures
