+++ pkg/services/internal/arc/arc_service_test.go
@@ -5,7 +5,8 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/4chain-ag/go-wallet-toolbox/pkg/services/internal/testabilities"
+	"github.com/4chain-ag/go-wallet-toolbox/pkg/internal/testabilities"
+	arctestabilities "github.com/4chain-ag/go-wallet-toolbox/pkg/services/internal/arc/testabilities"
 	"github.com/4chain-ag/go-wallet-toolbox/pkg/wdk"
 	sdk "github.com/bsv-blockchain/go-sdk/transaction"
 	txtestabilities "github.com/bsv-blockchain/universal-test-vectors/pkg/testabilities"
@@ -17,13 +18,13 @@ func TestPostBEEFWithARCService(t *testing.T) {
 
 	t.Run("broadcast without passing txid", func(t *testing.T) {
 		// given:
-		given := testabilities.Given(t)
+		given := arctestabilities.Given(t)
 
 		// setup arc server
 		given.ARC().IsUpAndRunning()
 
 		// and:
-		service := given.Services().NewArcService()
+		service := given.NewArcService()
 
 		// and:
 		tx := txtestabilities.GivenTX().WithInput(100).WithP2PKHOutput(99).TX()
@@ -52,13 +53,13 @@ func TestPostBEEFWithARCService(t *testing.T) {
 
 	t.Run("broadcast single transaction", func(t *testing.T) {
 		// given:
-		given := testabilities.Given(t)
+		given := arctestabilities.Given(t)
 
 		// setup arc server
 		given.ARC().IsUpAndRunning()
 
 		// and:
-		service := given.Services().NewArcService()
+		service := given.NewArcService()
 
 		// and:
 		tx := txtestabilities.GivenTX().WithInput(100).WithP2PKHOutput(99).TX()
@@ -88,13 +89,13 @@ func TestPostBEEFWithARCService(t *testing.T) {
 
 	t.Run("broadcast multiple txids", func(t *testing.T) {
 		// given:
-		given := testabilities.Given(t)
+		given := arctestabilities.Given(t)
 
 		// setup arc server
 		given.ARC().IsUpAndRunning()
 
 		// and:
-		service := given.Services().NewArcService()
+		service := given.NewArcService()
 
 		// and:
 		parentTx := txtestabilities.GivenTX().WithInput(100).WithP2PKHOutput(99).TX()
@@ -133,14 +134,14 @@ func TestPostBEEFWithARCService(t *testing.T) {
 
 	t.Run("return success if broadcast finished with OK without body, but we can query the tx", func(t *testing.T) {
 		// given:
-		given := testabilities.Given(t)
+		given := arctestabilities.Given(t)
 
 		// setup arc server
 		given.ARC().IsUpAndRunning()
 		given.ARC().OnBroadcast().WillReturnNoBody()
 
 		// and:
-		service := given.Services().NewArcService()
+		service := given.NewArcService()
 
 		// and:
 		tx := txtestabilities.GivenTX().WithInput(100).WithP2PKHOutput(99).TX()
@@ -210,13 +211,13 @@ func TestPostBEEFWithARCService(t *testing.T) {
 	for name, test := range invalidBEEFTestCases {
 		t.Run(name, func(t *testing.T) {
 			// given:
-			given := testabilities.Given(t)
+			given := arctestabilities.Given(t)
 
 			// setup arc server
 			given.ARC().IsUpAndRunning()
 
 			// and:
-			service := given.Services().NewArcService()
+			service := given.NewArcService()
 
 			// when:
 			res, err := service.PostBEEF(context.Background(), test.BEEF(t), nil)
@@ -257,13 +258,13 @@ func TestPostBEEFWithARCService(t *testing.T) {
 	for name, test := range arcFailingTestCases {
 		t.Run(name, func(t *testing.T) {
 			// given:
-			given := testabilities.Given(t)
+			given := arctestabilities.Given(t)
 
 			// setup arc server
 			test.setupARC(given.ARC())
 
 			// and:
-			service := given.Services().NewArcService()
+			service := given.NewArcService()
 
 			// and:
 			tx := txtestabilities.GivenTX().WithInput(100).WithP2PKHOutput(99).TX()
@@ -326,13 +327,13 @@ func TestPostBEEFWithARCService(t *testing.T) {
 	for name, test := range errorOnQueryTxTestCases {
 		t.Run(name, func(t *testing.T) {
 			// given:
-			given := testabilities.Given(t)
+			given := arctestabilities.Given(t)
 
 			// setup arc server
 			given.ARC().IsUpAndRunning()
 
 			// and:
-			service := given.Services().NewArcService()
+			service := given.NewArcService()
 
 			// and:
 			grandParentTx := txtestabilities.GivenTX().WithInput(300).WithP2PKHOutput(299).TX()
