+++ pkg/storage/internal/actions/process.go
@@ -191,7 +191,13 @@ func (p *process) broadcastSingleTx(ctx context.Context, txID string) (wdk.SendW
 
 	// TODO: SPV of the beef
 
-	_ = beef // TODO Services::PostBEEF
+	_ = beef
+	//results, err := p.services.PostBEEF(ctx, beef, []string{txID})
+	//if err != nil {
+	//	return "", fmt.Errorf("failed to post BEEF: %w", err)
+	//}
+	//
+	//_ = results // TODO postprocess results
 	return wdk.SendWithResultStatusSending, nil
 }
